Website Single cell
This contains the process required for the single cell UMAP visualization, differential, Gene scores. 

The differential can be performed for different samples, SARS-CoV-2 vaccines. 

The plots and table are easily downloadable using download button. 

It is very useful for people with no computation background, just like querying the genes or protein on the website.